tuto239 ondelete worked

+++ app/Http/Controllers/AdminUsersController.php
@@ -146,7 +146,7 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        unlink(public_path().$user->photo->file);
+        if($user->photo){ unlink(public_path().$user->photo->file);}
         $user->delete();
         Session::flash('user_deleted', 'the user has been deleted ');
         return redirect('/admin/users');
